⚡️ Improve code : Detach components to each file
