+++ src/List.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+function List({ onClick, toDos, key }) {
+  return (
+    <>
+      {toDos.map((toDo) => (
+        <li key={key}>
+          {toDo}
+          <button type="button" onClick={onClick}>
+            완료
+          </button>
+        </li>
+      ))}
+    </>
+  );
+}
+
+export default List;
