+++ src/index.jsx
@@ -1,46 +1,8 @@
 import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
 
-function Ul({ onClick, toDos, id }) {
-  return (
-    <ul>
-      {toDos[0] === undefined ? (<h3>할 일이 없어요!</h3>)
-        : (
-          <List
-            onClick={onClick}
-            toDos={toDos}
-            key={id}
-          />
-        )}
-    </ul>
-  );
-}
-
-function List({ onClick, toDos, key }) {
-  return (
-    <>
-      {toDos.map((toDo) => (
-        <li key={key}>
-          {toDo}
-          <button type="button" onClick={onClick}>
-            완료
-          </button>
-        </li>
-      ))}
-    </>
-  );
-}
-
-function InputForm({ onClick }) {
-  return (
-    <form>
-      <input id="input" type="text" required />
-      <button type="submit" onClick={onClick}>
-        추가
-      </button>
-    </form>
-  );
-}
+import InputForm from './InputForm';
+import UnOrderList from './UnOrderList';
 
 function App() {
   const [state, setState] = useState({
@@ -73,7 +35,7 @@ function App() {
     <div>
       <h1>To-do</h1>
       <InputForm onClick={handleClick} />
-      <Ul
+      <UnOrderList
         onClick={handleClickCompleteList}
         toDos={toDos}
         id={id}
